g6-common: Add blueprint-based Google Apps debloat
This works by adding a dummy .apk file (literally the default Android Studio project) that overrides all the packages you wish to remove.

Future changes will include building the .apk during the build process (instead of including a presigned and prebuilt file) in order to publish the source files for the dummy app.